requirements.txt
keystoneauth1==2.20.0
requests-futures==0.9.7
py-zabbix==1.1.3
xlsxwriter==1.0.0
request==0.0.13
urllib3==1.22
mime==0.1.0
requests-futures==0.9.7
email==6.0.0a1
Flask==0.12.2
Flask-RESTful==0.3.6
uWSGI==2.0.15